+++ .phan/config.php
@@ -452,10 +452,14 @@ return [
         'sysvshm'     => '.phan/internal_stubs/sysvshm.phan_php',
     ],
 
-    // Set this to false to emit PhanUndeclaredFunction issues for internal functions that Phan has signatures for,
-    // but aren't available in the codebase, or the internal functions used to run phan (may lead to false positives if an extension isn't loaded)
+    // Set this to false to emit `PhanUndeclaredFunction` issues for internal functions that Phan has signatures for,
+    // but aren't available in the codebase, or from Reflection.
+    // (may lead to false positives if an extension isn't loaded)
+    //
     // If this is true(default), then Phan will not warn.
-    // Also see 'autoload_internal_extension_signatures' for an alternative way to fix this type of issue.
+    //
+    // Even when this is false, Phan will still infer return values and check parameters of internal functions
+    // if Phan has the signatures.
     'ignore_undeclared_functions_with_known_signatures' => false,
 
     'plugin_config' => [